Diego branch

+++ client/src/Components/admin/AdminProducts/utils/request.js
@@ -17,28 +17,78 @@ export function sendChanges(product) {
     });
 }
 
-export function deleteProduct(product) {
-  axios.delete(`${URL_PRODUCTS}${product.id}`, product)
+export function createProduct(product, setLoading) {
+  axios.post(URL_PRODUCTS, product)
     .then((res) => {
       if (res.data.hasOwnProperty('err')) {
-        return swal('Error', res.data.err, 'warning');
+        setLoading(false);
+        swal('error', 'No se pudo crear al producto', 'warning');
+      } else {
+        setLoading(false);
+        swal('Success', '¡Se creó el producto con éxito!');
+        window.location.href = '/admin/products';
       }
-      return swal('Success', 'Producto eliminado con éxito!');
     })
     .catch(() => {
-      swal('Error', 'Ocurrió un error. No se pudo eliminar el producto. Intente nuevamente');
+      setLoading(false);
+      swal('Error', 'Ocurrió un error. No se creó el product, intenta nuevamente');
     });
 }
 
+export function deleteProduct(product) {
+  swal({
+    title: '¿Estás seguro que deseas eliminar este producto?',
+    text: '',
+    icon: 'warning',
+    buttons: true,
+    dangerMode: true,
+  }).then((willDelete) => {
+    if (willDelete) {
+      axios.delete(`${URL_PRODUCTS}${product.id}`, product)
+        .then((res) => {
+          if (res.data.hasOwnProperty('err')) {
+            swal('Error', res.data.err, 'warning');
+          } else {
+            swal('¡El producto ha sido eliminada con éxito!', {
+              icon: 'success',
+            });
+            window.location.href = '/admin/products';
+          }
+        })
+        .catch(() => {
+          swal('Error', 'Ocurrió un error. No se pudo eliminar el producto. Intente nuevamente');
+        });
+    } else {
+      swal('¡El producto no se eliminó!');
+    }
+  });
+}
+
 export function deleteCategory(category) {
-  axios.delete(`${URL_CATEGORIES}${category.id}`, category)
-    .then((res) => {
-      if (res.data.hasOwnProperty('err')) {
-        return swal('Error', res.data.err, 'warning');
-      }
-      return swal('Success', 'Producto eliminado con éxito!');
-    })
-    .catch(() => {
-      swal('Error', 'Ocurrió un error. No se pudo eliminar el producto. Intente nuevamente');
-    });
+  swal({
+    title: '¿Estás seguro que deseas eliminar esta categoría?',
+    text: '',
+    icon: 'warning',
+    buttons: true,
+    dangerMode: true,
+  }).then((willDelete) => {
+    if (willDelete) {
+      axios.delete(`${URL_CATEGORIES}${category.id}`, category)
+        .then((res) => {
+          if (res.data.hasOwnProperty('err')) {
+            swal('Error', res.data.err, 'warning');
+          } else {
+            swal('¡La categoría ha sido eliminada con éxito!', {
+              icon: 'success',
+            });
+            window.location.href = '/admin/controlcategories';
+          }
+        })
+        .catch(() => {
+          swal('Error', 'Ocurrió un error. No se pudo eliminar la categoría. Intente nuevamente');
+        });
+    } else {
+      swal('¡La categoría no se eliminó!');
+    }
+  });
 }
